Perinatal task review

+++ tablet_qt/tasks/ors.h
@@ -55,7 +55,7 @@ public:
     // Task specific
     // ------------------------------------------------------------------------
     void updateMandatory();
-    int totalScore() const;
+    double totalScore() const;
 public slots:
 
 public:
